Fix of doc tests

+++ pb_bss/evaluation/module_si_sdr.py
@@ -18,19 +18,19 @@ def si_sdr(reference, estimation):
     >>> np.random.seed(0)
     >>> reference = np.random.randn(100)
     >>> si_sdr(reference, reference)
-    inf
+    np.float64(inf)
     >>> si_sdr(reference, reference * 2)
-    inf
+    np.float64(inf)
     >>> si_sdr(reference, np.flip(reference))
-    -25.127672346460717
+    np.float64(-25.127672346460717)
     >>> si_sdr(reference, reference + np.flip(reference))
-    0.481070445785553
+    np.float64(0.481070445785553)
     >>> si_sdr(reference, reference + 0.5)
-    6.3704606032577304
+    np.float64(6.3704606032577304)
     >>> si_sdr(reference, reference * 2 + 1)
-    6.3704606032577304
+    np.float64(6.3704606032577304)
     >>> si_sdr([1., 0], [0., 0])  # never predict only zeros
-    nan
+    np.float64(nan)
     >>> si_sdr([reference, reference], [reference * 2 + 1, reference * 1 + 0.5])
     array([6.3704606, 6.3704606])
